+++ src/azure/graph/client.ts
@@ -1,6 +1,8 @@
 import {
   IntegrationLogger,
   IntegrationProviderAPIError,
+  IntegrationProviderAuthorizationError,
+  IntegrationValidationError,
 } from '@jupiterone/integration-sdk-core';
 import { FetchError } from 'node-fetch';
 import {
@@ -11,8 +13,9 @@ import {
 import { Organization } from '@microsoft/microsoft-graph-types';
 
 import { IntegrationConfig } from '../../types';
-import authenticate from './authenticate';
+import { authenticate } from './authenticate';
 import { retry } from '@lifeomic/attempt';
+import { permissions } from '../../steps/constants';
 
 interface AzureGraphResponse<TResponseType = any> {
   value: TResponseType[];
@@ -26,14 +29,72 @@ interface AzureGraphResponse<TResponseType = any> {
  */
 export abstract class GraphClient {
   protected client: Client;
+  private authenticationProvider: AuthenticationProvider;
 
   constructor(
     readonly logger: IntegrationLogger,
     readonly config: IntegrationConfig,
   ) {
+    this.authenticationProvider = new GraphAuthenticationProvider(config);
     this.client = Client.initWithMiddleware({
-      authProvider: new GraphAuthenticationProvider(config),
+      authProvider: this.authenticationProvider,
     });
+  }
+
+  private async getAccessTokenPermissions(): Promise<string[]> {
+    function getRolesFromAccessToken(accessToken: string) {
+      function parseJwtPayload(jwt: string): { roles?: string[] } | undefined {
+        try {
+          const encodedPayload = jwt.split('.')[1];
+          const decodedPayload = Buffer.from(
+            encodedPayload,
+            'base64',
+          ).toString();
+          return JSON.parse(decodedPayload);
+        } catch (e) {
+          return undefined;
+        }
+      }
+
+      const payload = parseJwtPayload(accessToken);
+      return payload?.roles || [];
+    }
+
+    const accessToken = await this.authenticationProvider.getAccessToken();
+    return getRolesFromAccessToken(accessToken);
+  }
+
+  protected async enforceApiPermission(
+    endpoint: string,
+    permission: string,
+  ): Promise<void> {
+    if (!(await this.getAccessTokenPermissions()).includes(permission)) {
+      throw new IntegrationProviderAuthorizationError({
+        endpoint,
+        status: 'MISSING_API_PERMISSION',
+        statusText: 'Missing API permission required to access this endpoint',
+        resourceType: [permission],
+      });
+    }
+  }
+
+  public async validate(): Promise<void> {
+    try {
+      await this.authenticationProvider.getAccessToken();
+    } catch (e) {
+      throw new IntegrationValidationError(e);
+    }
+  }
+
+  public async validateDirectoryPermissions(): Promise<void> {
+    try {
+      await this.enforceApiPermission(
+        '/organization',
+        permissions.graph.DIRECTORY_READ_ALL,
+      );
+    } catch (e) {
+      throw new IntegrationValidationError(e);
+    }
   }
 
   public async fetchMetadata(): Promise<AzureGraphResponse | undefined> {
@@ -104,7 +165,7 @@ export abstract class GraphClient {
           cause: err,
           endpoint,
           status: err.statusCode,
-          statusText: err.statusText,
+          statusText: err.statusText || err.message,
         });
       }