+++ src/steps/constants.ts
@@ -1,0 +1,7 @@
+export const permissions = {
+  graph: {
+    DIRECTORY_READ_ALL: 'Directory.Read.All',
+    POLICY_READ_ALL: 'Policy.Read.All',
+  },
+  resourceManager: {},
+};